Tracker and quantifier of wooded areas (green).
This code requires an image to define a scale factor and an urban area image taken from GoogleMaps. This software tracks and quantifies all wooded areas (green) in the image, finds the center of mass of the contours and the system and draws these contours in order to improve their identification, in addition to knowing the percentage of area tree (green) in the image.